fix: resolve logic constructs
